requirements.txt
dnspython==2.3.0
ipaddress==1.0.23
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
